+++ packages/core/src/llm/router/providers/grok.ts
@@ -1,11 +1,11 @@
 import { OpenAIChatSettings } from '@ai-sdk/openai/internal';
 import { createXai } from '@ai-sdk/xai';
 
-import { AISDK } from '../aisdk';
+import { ModelRouter } from '../router';
 
 export type XGrokModel = 'grok-beta' | 'grok-vision-beta' | 'grok-2-1212' | 'grok-2-vision-1212' | (string & {});
 
-export class Grok extends AISDK {
+export class Grok extends ModelRouter {
   constructor({
     name = 'grok-beta',
     apiKey = process.env.XAI_API_KEY ?? '',
