+++ packages/core/src/llm/router/providers/openai.test.ts
