+++ packages/core/src/llm/types.ts
@@ -31,39 +31,6 @@ export type CoreToolMessage = AiCoreToolMessage;
 export type EmbedResult<T> = AiEmbedResult<T>;
 
 export type EmbedManyResult<T> = AiEmbedManyResult<T>;
-export interface GoogleGenerativeAISettings {
-  /**
-  Optional.
-  The name of the cached content used as context to serve the prediction.
-  Format: cachedContents/{cachedContent}
-     */
-  cachedContent?: string;
-  /**
-   * Optional. Enable structured output. Default is true.
-   *
-   * This is useful when the JSON Schema contains elements that are
-   * not supported by the OpenAPI schema version that
-   * Google Generative AI uses. You can use this to disable
-   * structured outputs if you need to.
-   */
-  structuredOutputs?: boolean;
-  /**
-  Optional. A list of unique safety settings for blocking unsafe content.
-     */
-  safetySettings?: Array<{
-    category:
-    | 'HARM_CATEGORY_HATE_SPEECH'
-    | 'HARM_CATEGORY_DANGEROUS_CONTENT'
-    | 'HARM_CATEGORY_HARASSMENT'
-    | 'HARM_CATEGORY_SEXUALLY_EXPLICIT';
-    threshold:
-    | 'HARM_BLOCK_THRESHOLD_UNSPECIFIED'
-    | 'BLOCK_LOW_AND_ABOVE'
-    | 'BLOCK_MEDIUM_AND_ABOVE'
-    | 'BLOCK_ONLY_HIGH'
-    | 'BLOCK_NONE';
-  }>;
-}
 
 export type GenerateReturn<Z extends ZodSchema | JSONSchema7 | undefined = undefined> = Z extends undefined
   ? GenerateTextResult<any, any>
